requirements.txt
cffi==1.15.0
cryptography==36.0.2
greenlet==1.1.2
protobuf==3.19.4
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.32
